refactor: rename directory for dialogs
