+++ app/src/main/kotlin/view/dialogs/FordBellmanDialog.kt
@@ -1,0 +1,8 @@
+package view.dialogs
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun FordBellmanDialog() {
+
+}
